get data from order
