+++ src/containers/BurgerBuilder/BurgerBuilder.js
@@ -106,35 +106,11 @@ class BurgerBuilder extends Component {
 
     purchaseContinueHandler = () => {
         //alert('you continue')
-        /*
-        this.setState({ loading: true });
-        const order = {
-            ingredients: this.state.ingredients,
-            price: this.state.totalPrice,
-            customer: {
-                name: 'Max Shcwarxmuller',
-                address: {
-                    street: 'Teststreet 1',
-                    zipCode: '41351',
-                    country: 'Germany'
-                },
-                email: 'test@test.com'
-            },
-            deliveryMethod: 'fastest'
-        }
-        axios.post('/orders.json', order)
-            .then(response =>{
-                this.setState({ loading: false, purchasing: false });
-                console.log(response);
-            })
-            .catch(error =>{
-                this.setState({ loading: false, purchasing: false });
-            })
-    */
         const queryParams = [];
         for (let i in this.state.ingredients) {
             queryParams.push(encodeURIComponent(i) + '=' + encodeURIComponent(this.state.ingredients[i]));
         }
+        queryParams.push('price=' + this.state.totalPrice);
         const queryString = queryParams.join('&');    
         this.props.history.push({
             pathname: '/checkout',
