+++ src/containers/Checkout/Checkout.js
@@ -5,21 +5,22 @@ import ContactData from './ContactData/ContactData'
 
 class Checkout extends Component {
     state = {
-        ingredients: {
-            salad: 1,
-            meat: 1,
-            cheese: 1,
-            bacon: 1
-        }
+        ingredients: null,
+        price: 0
     }
 
-    componentDidMount () {
+    componentWillMount () {
         const query = new URLSearchParams(this.props.location.search);
         const ingredients = {};
+        let price = 0;
         for (let param of query.entries()) {
-            ingredients[param[0]] = +param[1];
+            if (param[0] === 'price') {
+                price = param[1]
+            } else {
+                ingredients[param[0]] = +param[1];
+            }
         }
-        this.setState({ingredients:ingredients});
+        this.setState({ingredients:ingredients, totalPrice:price});
     }
 
     checkoutCancelledHandler = () => {
@@ -39,7 +40,7 @@ class Checkout extends Component {
                     checkoutContinued={this.checkoutContinuedHandler} />
                 <Route 
                     path={this.props.match.path + '/contact-data'} 
-                    component={ContactData} />
+                    render={(props)=>(<ContactData ingredients={this.state.ingredients} price={this.totalPrice} {...props} />)} />
             </div>
         );
     }
