+++ src/containers/Checkout/ContactData/ContactData.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react'
 import Button from '../../../components/UI/Button/Button'
+import Spinner from '../../../components/UI/Spinner/Spinner'
 import classes from './ContactData.css'
-
+import axios from '../../../axios-orders'
 
 class ContactData extends Component {
     state = {
@@ -10,20 +11,54 @@ class ContactData extends Component {
         adress: {
             street: '',
             postalCode: ''
+        },
+        loading: false
+    }
+
+    orderHandler = (event) => {
+        event.preventDefault();
+        this.setState({ loading: true });
+        const order = {
+            ingredients: this.props.ingredients,
+            price: this.props.price,
+            customer: {
+                name: 'Max Shcwarxmuller',
+                address: {
+                    street: 'Teststreet 1',
+                    zipCode: '41351',
+                    country: 'Germany'
+                },
+                email: 'test@test.com'
+            },
+            deliveryMethod: 'fastest'
         }
+        axios.post('/orders.json', order)
+            .then(response =>{
+                this.setState({ loading: false });
+                this.props.history.push('/');
+            })
+            .catch(error =>{
+                this.setState({ loading: false });
+            })
     }
 
     render () {
+        let form = (
+            <form>
+                <input className={classes.Input} type="text" name="name" placeholder="Your Name" />
+                <input className={classes.Input} type="email" name="email" placeholder="Your Mail" />
+                <input className={classes.Input} type="text" name="street" placeholder="Street" />
+                <input className={classes.Input} type="text" name="postal" placeholder="Postal Code" />
+                <Button btnType="Success" clicked={this.orderHandler}>ORDER</Button>
+            </form>
+        );
+        if (this.state.loading) {
+            form = <Spinner />
+        }
         return (
             <div className={classes.ContactData}>
                 <h4>Enter your Contact Data</h4>
-                <form>
-                    <input className={classes.Input} type="text" name="name" placeholder="Your Name" />
-                    <input className={classes.Input} type="email" name="email" placeholder="Your Mail" />
-                    <input className={classes.Input} type="text" name="street" placeholder="Street" />
-                    <input className={classes.Input} type="text" name="postal" placeholder="Postal Code" />
-                    <Button btnType="Success">ORDER</Button>
-                </form>
+                {form}
             </div>
         );
     }
